bteb

+++ PHP/deleteEvenement.php
@@ -1,8 +1,12 @@
 <?php
 chdir("..");
-require_once 'DB/DB.inc.php';
 
-
+session_start();
+require_once "PHP/fonctions/func.seance.php";
+require_once("PHP/Twig/lib/Twig/Autoloader.php");
+require_once "PHP/fonctions/func.typeseance.php";
+require_once "PHP/fonctions/func.modules.php";
+require_once "PHP/fonctions/func.groupes.php";
 // TODO: à transformer en fonction
 
 
@@ -10,17 +14,30 @@ $db = DB::getInstance();
 if ($db == null) {
     echo "Impossible de se connecter &agrave; la base de donneeacute;es !";
 } else {
-    try {
-        $db->deleteEvenement($_GET['id']);
-        echo '<meta http-equiv="refresh" content="0;URL=../listEvenement.php">';
+    if(isset($_GET['id'])) {
 
-    } catch (Exception $e) {
-        echo "ça MARCHE PAS: " . $e->getMessage();
+    	$evenement = getEvenement($_GET['id'])[0];
+    	$idseance = $evenement->getIdSeance();
+    	$seance->getSeance($idseance)[0];
+    	echo $idseance;
+    	echo $seance;
+    	if( $seance->getIdSeance()->get === $_SESSION['login']) {
+    		
+    		try {
+        		$db->deleteEvenement($_GET['id']);
+        		echo '<meta http-equiv="refresh" content="0;URL=../listEvenement.php">';
 
-    }
+    		} catch (Exception $e) {
+        		echo "ça MARCHE PAS: " . $e->getMessage();
+
+    		}
+		}
+	}
+	echo 'test';
+	echo '<meta http-equiv="refresh" content="0;URL=../listEvenement.php">';
+
 }
-
-
 $db->close();
 
 
+
